Move data utils to utils/data (for now -- some of this is getting split off again into the store pending the history overhaul)
