+++ client/galaxy/scripts/utils/data.js
